Removed globals; Removed need for end_tests clause
